[NEW RBO] Inline projection pruning helpers
