--- ydb/core/kqp/opt/rbo/rules/map/projection_pruning_helpers.h
@@ -1,16 +1,0 @@
-#pragma once
-
-#include <ydb/core/kqp/opt/rbo/rules/kqp_rules_include.h>
-
-namespace NKikimr {
-namespace NKqp {
-
-TVector<TMapElement> KeepLiveMapElements(const TIntrusivePtr<TOpMap>& map, const TInfoUnitSet& liveOut, const TInfoUnitSet& keepKeyColumns = {});
-TVector<TInfoUnit> KeepLiveColumns(const TVector<TInfoUnit>& columns, const TInfoUnitSet& liveOut, const TInfoUnitSet& keepKeyColumns = {});
-TVector<TInfoUnit> BuildMapOutput(const TVector<TInfoUnit>& inputOutput, const TVector<TMapElement>& elements);
-TVector<TInfoUnit> BuildMapOutput(const TIntrusivePtr<TOpMap>& map, const TVector<TMapElement>& elements);
-bool NarrowReadColumns(const TIntrusivePtr<TOpRead>& read, const TVector<TInfoUnit>& liveOutput);
-bool PruneAggregateTraits(const TIntrusivePtr<TOpAggregate>& aggregate, const TVector<TInfoUnit>& liveOutput);
-
-} // namespace NKqp
-} // namespace NKikimr
