+++ ydb/core/kqp/opt/rbo/rules/map/push_rename_through_join_side.cpp
@@ -1,5 +1,5 @@
 #include <ydb/core/kqp/opt/rbo/rules/map/rename_common.h>
-#include <ydb/core/kqp/opt/rbo/rules/map/projection_pruning_helpers.h>
+#include <ydb/core/kqp/opt/rbo/rules/map/map_output_utils.h>
 
 namespace NKikimr {
 namespace NKqp {
