+++ ydb/core/kqp/opt/rbo/rules/ya.make
@@ -19,8 +19,8 @@ SRCS(
     inline_generic_in_exists_subplan.cpp
     kqp_cbo_trees.cpp
     peephole_predicate.cpp
+    map/map_output_utils.cpp
     map/projection_pruning.cpp
-    map/projection_pruning_helpers.cpp
     map/push_append_into_map.cpp
     map/push_append_through_aggregate.cpp
     map/push_append_through_join.cpp
